调试版本 i2c从机无响应 uprintf只能打印有符号整型 待继续修改
测试I2C接口，tuner从机不响应I2C

+++ io.c
@@ -25,7 +25,7 @@ char uprintf(const char *fmt, ...)
                 break;
             case 'u':   
                 {
-                    d = va_arg(ap,long int);   //0~4294967295
+                    d = va_arg(ap,unsigned int);   //0~4294967295
                     i=0;        //turn d into decimal,low digit first
                     while(1)
                       {
